requirements/base.txt
Django==4.1.7
environs[django]==9.5.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-cors-headers==3.14.0
PyYAML==6.0
celery==5.2.7
redis==4.5.3
requests==2.28.2
django-celery-beat==2.5.0